+++ src/main/java/org/intellimate/izou/security/IzouSecurityManager.java
@@ -83,9 +83,6 @@ public class IzouSecurityManager extends SecurityManager {
      */
     private void init() {
         String workingDir = FileSystemManager.FULL_WORKING_DIRECTORY;
-        while (!workingDir.endsWith(File.separator + "Izou" + File.separator)) {
-            workingDir = workingDir.substring(0, workingDir.length() - 1);
-        }
 
         forbiddenProperties.add("jdk.lang.process.launchmechanism");
